Add confirm prompt for delete. Should be fancied if an opportunity presents

+++ src/frontend/components/admin/DeleteGroupButton.jsx
@@ -4,7 +4,10 @@ import React from 'react';
 export default ({ onDelete, group }) => {
 
     var deleteGroup = () => {
-        onDelete(group);
+        if(confirm('Are you sure you want to delete the current group?')){
+            onDelete(group);
+        }
     };
+    
     return (<button onClick={deleteGroup} className="deleteGroup" title="Delete group"><span>Delete group</span></button>)
 }
